[After applying this commit it will...] Move classes to new packages
