--- games
@@ -1,4 +1,0 @@
-Tytuł gry: Grand Theft Auto 6
-Data premiery: 2025
-Wydawca: Rockstar Games
-Gatunek: ACTION
