+++ src/main/java/pl/pabianczyklukasz/project/domain/Client.java
@@ -1,11 +1,10 @@
-package pl.pabianczyklukasz.project;
+package pl.pabianczyklukasz.project.domain;
 
+import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
-
-import java.util.List;
 
 @ToString
 @AllArgsConstructor
