+++ src/main/java/pl/pabianczyklukasz/project/domain/Employee.java
@@ -1,8 +1,8 @@
-package pl.pabianczyklukasz.project;
-
-import lombok.AllArgsConstructor;
+package pl.pabianczyklukasz.project.domain;
 
 import java.util.List;
+import lombok.AllArgsConstructor;
+import pl.pabianczyklukasz.project.repositories.GameRepository;
 
 @AllArgsConstructor
