+++ src/main/java/pl/pabianczyklukasz/project/domain/TypeOfGame.java
@@ -1,4 +1,4 @@
-package pl.pabianczyklukasz.project;
+package pl.pabianczyklukasz.project.domain;
 
 public enum TypeOfGame {
     STRATEGIC,
