--- src/main/java/pl/pabianczyklukasz/project/games.csv
@@ -1,2 +1,0 @@
-Game(title=Grand Theft Auto 6, yearOfRelease=2025, publisher=Publisher(name=Rockstar Games), typeOfGame=ACTION)
-Game(title=Sims, yearOfRelease=2020, publisher=Publisher(name=EA), typeOfGame=SIMULATION)
