+++ src/main/java/pl/pabianczyklukasz/project/repositories/GameRepository.java
@@ -1,7 +1,8 @@
-package pl.pabianczyklukasz.project;
+package pl.pabianczyklukasz.project.repositories;
 
 import java.io.IOException;
-import java.util.List;
+import pl.pabianczyklukasz.project.domain.Game;
+import pl.pabianczyklukasz.project.domain.Publisher;
 
 public interface GameRepository {
     public void readAllGames();
